chore: Move more integration GraphQL types to SDL
There were some type resolver issues caused by #10000 and converting
more types to SDL was the easiest way to fix the issues caused by not
being able to resolve some interface types.
