+++ packages/server/graphql/public/typeDefs/JiraRemoteProject.graphql
@@ -1,0 +1,30 @@
+"""
+A project fetched from Jira in real time
+"""
+type JiraRemoteProject implements RepoIntegration {
+  id: ID!
+  service: IntegrationProviderServiceEnum!
+
+  """
+  The parabol teamId this issue was fetched for
+  """
+  teamId: ID!
+
+  """
+  The parabol userId this issue was fetched for
+  """
+  userId: ID!
+  self: ID!
+
+  """
+  The cloud ID that the project lives on. Does not exist on the Jira object!
+  """
+  cloudId: ID!
+  key: String!
+  name: String!
+  avatar: String
+  avatarUrls: JiraRemoteAvatarUrls!
+  projectCategory: JiraRemoteProjectCategory!
+  simplified: Boolean!
+  style: String!
+}
