+++ packages/server/graphql/public/typeDefs/JiraServerRemoteProject.graphql
@@ -1,0 +1,21 @@
+"""
+A project fetched from Jira in real time
+"""
+type JiraServerRemoteProject implements RepoIntegration {
+  id: ID!
+  service: IntegrationProviderServiceEnum!
+
+  """
+  The parabol teamId this issue was fetched for
+  """
+  teamId: ID!
+
+  """
+  The parabol userId this issue was fetched for
+  """
+  userId: ID!
+  name: String!
+  avatar: String!
+  avatarUrls: JiraRemoteAvatarUrls!
+  projectCategory: JiraRemoteProjectCategory!
+}
