+++ packages/server/graphql/public/typeDefs/RepoIntegration.graphql
@@ -1,0 +1,7 @@
+"""
+The suggested repos and projects a user can integrate with
+"""
+interface RepoIntegration {
+  id: ID!
+  service: IntegrationProviderServiceEnum!
+}
