+++ packages/server/graphql/public/types/AddTeamMemberIntegrationAuthSuccess.ts
@@ -1,0 +1,23 @@
+import toTeamMemberId from '../../../../client/utils/relay/toTeamMemberId'
+import {AddTeamMemberIntegrationAuthSuccessResolvers} from '../resolverTypes'
+
+export type AddTeamMemberIntegrationAuthSuccessSource = {
+  service: string
+  teamId: string
+  userId: string
+}
+
+const AddTeamMemberIntegrationAuthSuccess: AddTeamMemberIntegrationAuthSuccessResolvers = {
+  integrationAuth: async ({service, teamId, userId}, _args, {dataLoader}) => {
+    return (await dataLoader.get('teamMemberIntegrationAuths').load({service, teamId, userId}))!
+  },
+  teamMember: ({teamId, userId}, _args, {dataLoader}) => {
+    const teamMemberId = toTeamMemberId(teamId, userId)
+    return dataLoader.get('teamMembers').load(teamMemberId)
+  },
+  user: async ({userId}, _args, {dataLoader}) => {
+    return dataLoader.get('users').loadNonNull(userId)
+  }
+}
+
+export default AddTeamMemberIntegrationAuthSuccess
